requirements.txt
chardet
pandas
python-dotenv
supabase
